chore: extract types to separate file
