+++ packages/preact/src/hooks/useSwatches/useSwatches.ts
@@ -46,9 +46,8 @@ import { filterSwatches } from "./filterSwatches"
 export function useSwatches(skus: SearchProductSku[] = [], fields: string[] = []) {
   const [selectedOptions, setSelectedOptions] = useState<Record<string, string>>({})
 
-  const filteredSwatches = useMemo(() => {
-    const swatches = aggregateSwatches(skus, fields)
-    return filterSwatches(swatches, selectedOptions)
+  const swatches = useMemo(() => {
+    return filterSwatches(aggregateSwatches(skus, fields), selectedOptions)
   }, [skus, fields, selectedOptions])
 
   const toggleOption = useCallback((field: string, value: string) => {
@@ -63,5 +62,5 @@ export function useSwatches(skus: SearchProductSku[] = [], fields: string[] = []
     })
   }, [])
 
-  return { swatches: filteredSwatches, toggleOption }
+  return { swatches, toggleOption }
 }
